requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
future==0.18.3
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
peewee==3.16.3
pefile==2021.9.3
pewee==0.1.0.dev9
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
yarl==1.9.2
